Module viaduct::backend::ffi

Modules

callback_holder
Module that manages get/set of the global fetch callback pointer.

Structs

FfiBackend

Functions

viaduct_alloc_bytebuffer
Return a ByteBuffer of the requested size. This is used to store the response from the callback.

viaduct_destroy_bytebuffer

viaduct_initialize

viaduct_log_error

Type Definitions

FetchCallback
Type of the callback we need callers on the other side of the FFI to provide.