viaduct

Module ohttp

Structs

OhttpConfig

Configuration for an OHTTP channel

Functions

clear_ohttp_channels

Clear all OHTTP channel configurations

configure_default_ohttp_channels

Configure default OHTTP channels for common Mozilla services This sets up:

configure_ohttp_channel

Configure an OHTTP channel with the given configuration If an existing OHTTP config exists with the same name, it will be overwritten

fetch_gateway_config

Fetch and cache gateway configuration (encryption keys)

get_ohttp_config

Get the configuration for a specific OHTTP channel

is_ohttp_channel_configured

Check if an OHTTP channel is configured

list_ohttp_channels

List all configured OHTTP channels

process_ohttp_request

Process an OHTTP request using the OHTTP client component