viaduct

Module settings

Structs

Settings

Note: reqwest allows these only to be specified per-Client. concept-fetch allows these to be specified on each call to fetch. I think it’s worth keeping a single global reqwest::Client in the reqwest backend, to simplify the way we abstract away from these.

Statics

GLOBAL_SETTINGS

Functions

allow_android_emulator_loopback

Allow non-HTTPS requests to the emulator loopback URL

set_global_default_user_agent

Set the global default user-agent

validate_request

Validate a request, respecting the addn_allowed_insecure_url setting.