viaduct

Struct Response

pub struct Response {
    pub request_method: Method,
    pub url: Url,
    pub status: u16,
    pub headers: Headers,
    pub body: Vec<u8>,
}

A response from the server.

Fields
request_method: Method

The method used to request this response.

url: Url

The URL of this response.

status: u16

The HTTP Status code of this response.

headers: Headers

The headers returned with this response.

body: Vec<u8>

The body of the response.

Implementations

impl Response

pub fn json<'a, T>(&'a self) -> Result<T, Error>
where T: Deserialize<'a>,

Parse the body as JSON.

pub fn text(&self) -> Cow<'_, str>

Get the body as a string. Assumes UTF-8 encoding. Any non-utf8 bytes are replaced with the replacement character.

pub fn is_success(&self) -> bool

Returns true if the status code is in the interval [200, 300).

pub fn is_server_error(&self) -> bool

Returns true if the status code is in the interval [500, 600).

pub fn is_client_error(&self) -> bool

Returns true if the status code is in the interval [400, 500).

pub fn require_success(self) -> Result<Self, UnexpectedStatus>

Returns an UnexpectedStatus error if self.is_success() is false, otherwise returns Ok(self).